fix: replace floats with doubles

+++ test/src/test_integration.cpp
@@ -102,7 +102,7 @@ int main() {
         test_network.set_biases(bias);
         double diff_sum = 0.0;
         for (int i = 0; i < bits_arr_sz; i++) {
-            float* network_out = test_network(bits_arr[i]);
+            double* network_out = test_network(bits_arr[i]);
             diff_sum += fabs(double(network_out[0]) - double(rule(bits_arr[i])));
             delete[] network_out;
         }
@@ -116,7 +116,7 @@ int main() {
     //m_dim_anneal.show_plot();
     std::cout << energy_fn(result) << std::endl;
     for (int i = 0; i < bits_arr_sz; i++) {
-        float* network_out = test_network(bits_arr[i]);
+        double* network_out = test_network(bits_arr[i]);
         std::cout << network_out[0] << " " << rule(bits_arr[i]) << std::endl;
         delete[] network_out;
     }
